backend/requirements.txt
Django>=5.2.5
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
psycopg2-binary>=2.9.10
python-dotenv>=1.1.1
openai>=1.99.9
django-cors-headers>=4.7.0
